requirements.txt
setuptools >= 21.0.0
boto3==1.18.50
requests==2.26.0
openai~=0.28.1

Pillow~=8.3.1
